backend/requirements.txt
blinker==1.8.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
freetype-py==2.4.0
future==1.0.0
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.22.0
msgpack==1.1.0
packaging==24.1
pillow==10.3.0
pycairo==1.26.1
pycparser==2.22
pylabels==1.2.1
pyzmq==26.2.1
reportlab==4.2.2
requests==2.32.3
setuptools==76.0.0
urllib3==2.3.0
Werkzeug==3.0.3
zope.event==5.0
zope.interface==7.2
